build(circular-deps): tsPreCompilationDeps only fix for video player
